Exception Handling

+++ professional/dscommerce/src/main/java/com/github/JBreno/dscommerce/services/ProductService.java
@@ -3,11 +3,14 @@ package com.github.JBreno.dscommerce.services;
 import com.github.JBreno.dscommerce.dto.ProductDTO;
 import com.github.JBreno.dscommerce.entities.Product;
 import com.github.JBreno.dscommerce.repositories.ProductRepository;
+import com.github.JBreno.dscommerce.services.exceptions.DatabaseException;
 import com.github.JBreno.dscommerce.services.exceptions.ResouceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -44,15 +47,29 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
-        Product entity = repository.getReferenceById(id);
-        copyDtoToEntity(dto, entity);
-        entity = repository.save(entity);
-        return new ProductDTO(entity);
+        try {
+            Product entity = repository.getReferenceById(id);
+            copyDtoToEntity(dto, entity);
+            entity = repository.save(entity);
+            return new ProductDTO(entity);
+        }
+        catch (Exception e) {
+            throw new ResouceNotFoundException("Resource not found");
+        }
+
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if(!repository.existsById(id)) {
+            throw new ResouceNotFoundException("Resource not found");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
